Feat: Napier Integration (#393)
* feat(napier): interface and utils

* chore: v4-core

* constant

* feat(fuse): napier

* feat: add Napier market constant

* feat(oracle): napier price oracle

* utils

* T

* feat: NapierPriceFeedFactory

* T

* refactor(napier): rename error and improve code style

- rename NapierFuseIInvalidMarketId to NapierFuseIInvalidToken for clarity

- extract principalToken to local variable to reduce storage reads

- remove unused ERC20 import

* test(napier): update revert error selector names

- align test expectations with renamed error from NapierFuseIInvalidMarketId to NapierFuseIInvalidToken

* feat(napier): add token validation for input and output tokens

- Add substrate asset grant checks for tokenIn/tokenOut in all fuse operations

- Prevent unauthorized tokens from being used in combine, redeem, supply, and swap operations

* test(napier): add token validation revert tests

- Add tests verifying supply/redeem revert when token not granted

- Add tests verifying swap PT/YT enter/exit revert for invalid tokens

- Update setup to grant underlying and base asset as substrates

- Add helper to selectively omit substrates for negative testing

* fix(napier): reset permit2 approval

* fix(oracle): Napier Price Feed

* T

* fix(oracle): PriceFeed 18 decimals

* refactor: Permit2 util

* refactor

* fix: validate rewards

* T

* feat(napier): add deposit and zap deposit fuses

- add NapierDepositFuse for proportional liquidity add/remove

- add NapierZapDepositFuse for single-sided currency1 deposits

- add LibExpiry utility for PT maturity checks

* T

* T: setup TWAP oracle

* T

* edit comments

* ✨(napier): add README documentation for Napier fuse integration

- Introduce comprehensive documentation outlining the Napier V2 yield-stripping protocol, including market structure, terminology, architecture, and operational details.
- Provide insights into key components and lifecycle management for PT and YT tokens, enhancing clarity for future development and integration.

* ✨(napier): implement NapierYtPriceFeed and ITokiOracle interfaces

- Introduce NapierYtPriceFeed contract for providing USD price data for Napier yield tokens using TWAP from TokiOracle.
- Add ITokiOracle interface for conversion functions and TWAP readiness checks.
- Ensure proper error handling and validation for oracle readiness and pricing.

* ✨(napier): enhance NapierPriceFeedFactory with YT price feed creation

- Add functionality to create Napier YT price feeds, including a new event for tracking their creation.
- Implement validation for input addresses to prevent zero address errors.
- Update the factory contract to support the new NapierYtPriceFeed integration.

* rename

* T

* rename NapierPriceFeed

* refactor(napier): moved ITokiChainlinkCompatOracle interface to a new file for better organization.
- Updated NapierPtLpPriceFeed to import the interface from its new location, enhancing code clarity and maintainability.
- Adjusted test imports accordingly to reflect the new structure.

* ✨(napier): introduce NapierYtLinearPriceFeed contract

* T

* refactor: rename  createPriceFeed -> createPtLpPriceFeed, createYtPriceFeed -> createYtTwapPriceFeed

* ✨(napier): add createYtLinearPriceFeed function to NapierPriceFeedFactory

- Introduced a new function to create Napier YT linear price feeds.
- Added an event to emit upon creation of the YT linear price feed.
- Implemented address validation to prevent zero address errors.

* refactor: fix typo

* T

* test(napier): correct PriceOracleMiddleware address

* test(napier): add Gauntlet USDC price feed

* style

* fix(price_feed):  handle post-expiry logic, returning a small price to bypass validation.

* T

* edit comments

* per(napier): cache array length

* feat(napier): add minimum output checks for Combine, Redeem, and Supply Fuses

* T

* feat(napier): validate amountIn > 0

* T

* refactor(price_feed): remove priceMiddleware_ parameter from price feed constructors and functions

* T

* docs

* fix(fuse): add slippage protection

* T

* typo

* typo

* bump solidity version

* style: lint:fix

+++ test/fuses/tac/TacValidatorAddressConverterTest.sol
@@ -21,7 +21,7 @@ contract TacValidatorAddressConverterTest is Test {
 
     function stestRoundTrip32Bytes() public {
         string memory input = string(abi.encodePacked(new bytes(32)));
-        for (uint i = 0; i < 32; i++) {
+        for (uint256 i = 0; i < 32; i++) {
             bytes(input)[i] = bytes1(uint8(65 + i)); // A, B, C, ...
         }
         (bytes32 a, bytes32 b) = TacValidatorAddressConverter.validatorAddressToBytes32(input);
@@ -31,7 +31,7 @@ contract TacValidatorAddressConverterTest is Test {
 
     function stestRoundTrip33Bytes() public {
         string memory input = string(abi.encodePacked(new bytes(33)));
-        for (uint i = 0; i < 33; i++) {
+        for (uint256 i = 0; i < 33; i++) {
             bytes(input)[i] = bytes1(uint8(97 + i)); // a, b, c, ...
         }
         (bytes32 a, bytes32 b) = TacValidatorAddressConverter.validatorAddressToBytes32(input);
@@ -41,7 +41,7 @@ contract TacValidatorAddressConverterTest is Test {
 
     function stestRoundTrip63Bytes() public {
         string memory input = string(abi.encodePacked(new bytes(63)));
-        for (uint i = 0; i < 63; i++) {
+        for (uint256 i = 0; i < 63; i++) {
             bytes(input)[i] = bytes1(uint8(33 + i));
         }
         (bytes32 a, bytes32 b) = TacValidatorAddressConverter.validatorAddressToBytes32(input);
@@ -51,7 +51,7 @@ contract TacValidatorAddressConverterTest is Test {
 
     function stestRoundTrip64Bytes() public {
         string memory input = string(abi.encodePacked(new bytes(63)));
-        for (uint i = 0; i < 63; i++) {
+        for (uint256 i = 0; i < 63; i++) {
             bytes(input)[i] = bytes1(uint8(128 + i));
         }
         (bytes32 a, bytes32 b) = TacValidatorAddressConverter.validatorAddressToBytes32(input);
